10월 31일 중간 작업
EnemyScript, WindSlashScript, BulletScript, PlayerStatus 등 스크립트 변화줌

데이터테이블의 데이터를 리스트화 시켜서 관리중

+++ Assets/10.DataTable/DB_StatusEntity.cs
@@ -77,7 +77,7 @@ public class DB_StatusEntity_03
 
     public int weaponType; // 기본공격 속성(타입)
 
-    public float typeLevel; // 기본공격 레벨
+    public int typeLevel; // 기본공격 레벨
 
     public string name; // 기본공격 이름
